code/requirements.txt
jupyter>=1.0.0
numpy>=1.11.1
scipy>=0.17.1
matplotlib>=1.5.0
easydict>=1.6
pillow>=3.2.0
scikit-learn>=0.17
scikit-image>=0.11.3
pandas>=0.17.1
colored>=1.2.2
h5py>=2.5.0
hickle>=2.0.5
Cython>=0.23.4
nltk>=3.2
tqdm>=4.8.2
tensorflow-gpu-1.2.1
spacy
